Create LoopAveragePositivNumber
We want to design a program that calculates the average of
of positive integers entered from the keyboard.

The user enters positive numbers on the keyboard, then to indicate the end of the
a negative number (condition for ending the input). The computer
displays the average of the numbers entered in the console.