Create BuildingPyramid
In one town, the inhabitants decide to build
a pyramid. They want to know how many storeys
they can build with the number of stones
they have.

The program asks you to enter the number of stones
stones available, and displays the number of floors
that can be built with them.